% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{greta.template}
\alias{greta.template}
\alias{greta.template-package}
\title{What the Package Does (One Line, Title Case)}
\description{
describe your package here, you can re-use the text from DESCRIPTION
}
\examples{

# add a simple example here to introduce the package!

}
