% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_check.greta_array.r
\name{pp_check.greta_array}
\alias{pp_check.greta_array}
\alias{pp_check}
\title{Prior Predictive Checks for \code{greta_array} object}
\usage{
\method{pp_check}{greta_array}(y, nsim = 100, ndraw = 10, ...)
}
\arguments{
\item{y}{A greta_array object generated from the greta::distributions function}

\item{nsim}{A numeric, the number of prior simulations per draw, default is 100}

\item{ndraw}{A numeric, the number of draws, default is 10}

\item{...}{Further arguments passed to bayesplot::ppc_dens_overlay, check the possibilities there.}
}
\value{
A ggplot object that can be further modified using the \pkg{ggplot2} package.
}
\description{
Perform prior predictive checks with the help of the package \code{bayesplot}
}
\examples{
\dontrun{
x <- normal(0, 2)

# default use
pp_check(x)
}

}
