% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_check.greta_mcmc_list.r
\name{pp_check.greta_mcmc_list}
\alias{pp_check.greta_mcmc_list}
\alias{pp_check}
\title{Posterior Predictive Checks for \code{greta_mcmc_list} object}
\usage{
\method{pp_check}{greta_mcmc_list}(object, y, nsim = 10, type = "dens_overlay", ...)
}
\arguments{
\item{object}{A greta_mcmc_list object obtained from the greta::mcmc function}

\item{y}{A greta array of the response variable, see example}

\item{nsim}{A numeric, the number of posterior simulation to draw, default is 10}

\item{type}{A character string, the type of posterior predictive plot to draw. See \code{\link[bayesplot:PPC-overview]{PPC-overview}} for some options, or check the example below.}

\item{...}{Further arguments passed to ppc_* functions, check the possibilities there.}
}
\value{
A ggplot object that can be further modified using the \pkg{ggplot2} package.
}
\description{
Perform posterior predictive checks with the help of the package \code{bayesplot}
}
\details{
For a detailed explanation of each of the implemented ppc functions, see \code{\link[bayesplot:PPC-overview]{PPC-overview}}.
Note that LOO and Discrete ppc_* function are not yet available for greta models.
}
\examples{
\dontrun{
x <- runif(100, -2, 2)
y <- rnorm(100, 1 + 2 * x, 1)
# need the pass the response vector as greta array
y <- as_data(y)

intercept <- normal(0, 2)
slope <- normal(0, 1)
linpred <- intercept + slope * x
sd_res <- cauchy(0, 2, truncation = c(0, Inf))

distribution(y) <- normal(linpred, sd_res)
m <- model(intercept, slope, sd_res)
d <- mcmc(m, warmup = 10, n_samples = 10)

# default use
pp_check(d, y)
# check some options
pp_check(d, y, nsim = 4, type = "scatter_avg")
pp_check(d, y, type = "hist")
pp_check(d, y, type = "error_scatter") 
}

}
