% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.r
\name{rmse}
\alias{rmse}
\title{compute the (normalized) root mean squared error for a greta regression model}
\usage{
rmse(y, pred, draws, summary = TRUE, probs = c(0.1, 0.9), norm = FALSE)
}
\arguments{
\item{y}{a greta array, the response variables}

\item{pred}{a greta array, the linear predictor}

\item{draws}{a greta_mcmc_list object, posterior draws as returned from calling greta sampling algorithm (ie mcmc)}

\item{summary}{a logical, if TRUE (default) the function output summary statistics (mean, sd, 80\% credible intervals) for the R2, if FALSE the raw values are returned}

\item{probs}{a vector of two numeric specifying the lower and upper limits for the credible intervals (default to 0.1, 0.9), only used if summary=TRUE}

\item{norm}{a logical, whether to normalize the RMSE by the mean of the response variable}
}
\value{
If summary=TRUE a 1 x C matrix is returned (C = length(probs) + 2) containing summary statistics of Bayesian R-squared values. If summary = FALSE the posterior samples of the R-squared values are returned as a numeric vector of length S (S is the number of samples)
}
\description{
Compute the root mean squared error of a greta model
}
\details{
Note that when normalized RMSE (norm=TRUE), is compute as follow: nRMSE = RMSE / mean(y)
}
\examples{
\dontrun{
intercept <- normal(0, 1)
slope <- normal(0, 1)
sd_resid <- cauchy(0, 1, truncation = c(0, 100))

x <- runif(100)
y <- as_data(rnorm(100, 1 + 2 * x, 1))

pred <- intercept + slope * x
distribution(y) <- normal(pred, sd_resid)

m <- model(intercept, slope, sd_resid)
drr <- mcmc(m)

rmse(y, pred, drr)
}
}
