% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{compute the square of a greta array}
\usage{
square(x)
}
\arguments{
\item{x}{a greta array}
}
\value{
a greta array
}
\description{
compute the square of a greta array
}
\details{
computes the square of x, in a slightly more efficient manner than
\code{x ^ 2}
}
\examples{
x <- variable(dim = c(3, 3))
y <- square(x)
}
