% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_check.greta_mcmc_list.r
\name{pp_check.greta_mcmc_list}
\alias{pp_check.greta_mcmc_list}
\alias{pp_check}
\title{Posterior Predictive Checks for \code{greta_mcmc_list} object}
\usage{
\method{pp_check}{greta_mcmc_list}(object, y, nsim = 10, type = "dens_overlay", ...)
}
\arguments{
\item{object}{A greta_mcmc_list object obtained from the greta::mcmc function}

\item{y}{A greta array of the response variable, see example}

\item{nsim}{A numeric, the number of posterior simulation to draw, default is 10}

\item{type}{A character string, the type of posterior predictive plot to draw. See \code{\link[bayesplot:PPC-overview]{PPC-overview}} for some options, or check the example below.}

\item{...}{Further arguments passed to ppc_* functions, check the possibilities there.}
}
\value{
A ggplot object that can be further modified using the \pkg{ggplot2} package.
}
\description{
Perform posterior predictive checks with the help of the package \code{bayesplot}
}
\details{
For a detailed explanation of each of the implemented ppc functions, see \code{\link[bayesplot:PPC-overview]{PPC-overview}}. Note that LOO and Discrete ppc_* function are not yet available for greta models.
}
